Rename testharness to apitest
