+++ api/apitest/package.go
@@ -12,4 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package testharness // import "go.opentelemetry.io/otel/api/testharness"
+package apitest // import "go.opentelemetry.io/otel/api/apitest"
